WIP file management WIP PCI WIP fixes for UnFS

+++ src/drivers/disk.rs
@@ -176,7 +176,7 @@ impl Disk {
     }
 
     //TODO: Fix and make sure count is not zero!
-    pub unsafe fn write(&self, lba: u64, count: u16, source: u32) -> u8{
+    pub unsafe fn write(&self, lba: u64, count: u16, source: usize) -> u8{
         while self.ide_read(ATA_REG_STATUS) & ATA_SR_BSY == ATA_SR_BSY {
 
         }
@@ -195,7 +195,7 @@ impl Disk {
 
         self.ide_write(ATA_REG_COMMAND, ATA_CMD_WRITE_PIO_EXT);
 
-        for sector in 0..count as u32 {
+        for sector in 0..count as usize {
             let err = self.ide_poll(true);
             if err > 0{
                 return err;
